feat(ci): Simplify CI and introduce devbox ( fixes: #32 )
This commit simplifies the CI by removing the Docker-based build and
test process. Instead, it introduces devbox, a tool that provides a
consistent development environment across different machines.

The CI now uses actions/checkout@v4 and devbox for setting up the
environment. The build and test commands are run using devbox.

The Dockerfile has been removed as it's no longer needed. The devbox
configuration is provided in devbox.json and devbox.lock.

The deployment to GitHub Pages has been commented out for now.